+++ src/locales/index.ts
@@ -1,7 +1,8 @@
 import type { App } from 'vue'
 import { createI18n } from 'vue-i18n'
-import en from './en-US'
-import cn from './zh-CN'
+import enUS from './en-US'
+import zhCN from './zh-CN'
+import zhTW from './zh-TW'
 import { useAppStoreWithOut } from '@/store/modules/app'
 import type { Language } from '@/store/modules/app/helper'
 
@@ -14,8 +15,9 @@ const i18n = createI18n({
   fallbackLocale: 'en-US',
   allowComposition: true,
   messages: {
-    'en-US': en,
-    'zh-CN': cn,
+    'en-US': enUS,
+    'zh-CN': zhCN,
+    'zh-TW': zhTW,
   },
 })